server/requirements.txt
fastapi
uvicorn
requests
qdrant-client
openai
PyJWT
python-multipart
pypdf
langchain_community
langchain_openai